\name{kftrack-internal}
\title{Internal kftrack objects}
\alias{.addrange}
\alias{.CI.reg}
\alias{.First}
\alias{.generate.dat.file}
\alias{.generate.twoseg.dat.file}
\alias{.gmtok}
\alias{.have.date}
\alias{.have.netCDF}
\alias{.init.scan}
\alias{.middate}
\alias{.plot1}
\alias{.plot2}
\alias{.plot3}
\alias{.read.output}
\alias{.read.twoseg.output}
\alias{.add.fit.to.server.list}
\alias{.system}
\description{Internal kftrack objects.}
\details{These are not to be called by the user.}
\keyword{internal}

