\name{kftrack-package}
\alias{kftrack-package}
\docType{package}
\title{
KFtrack estimates most probable track from archival tagging data
}
\description{
Fit a state space model to observed geo--locations via the extended Kalman filter.
A few variations of the model is available.    
}

\author{John Sibert \email{sibert@hawaii.edu}, Anders Nielsen \email{anders.nielsen@hawaii.edu}}

\references{Sibert, J., Musyl, M. K. and Brill, R.W. (2003) Horizontal movements
            of bigeye tuna near Hawaii determined by Kalman filter analysis of
	    archival tagging data. Fish. Oceanogr. 12(3):141--151.}

\keyword{models}

