\name{plot.kftrack.scan}
\alias{plot.kftrack.scan}
\title{Plot a kftrack scanning object}
\description{
Plots the result of a scanning for and estimation
of a premature pop--off position.
}
\usage{
plot.kftrack.scan(x, ...)
}
\arguments{
  \item{x}{is a \code{kftrack.scan} object typically generated
           with the \code{.init.scan} function.}
  \item{\dots}{additional graphics parameters}
}
\value{
  No value is returned this function is invoked for its side effects.
}

\author{John Sibert \email{jsibert@soest.hawaii.edu}, Anders Nielsen \email{anielsen@dina.kvl.dk}}

\seealso{\code{\link{kftrack}}, \code{\link{addmap}}, \code{\link{addcoast}}}

\keyword{programming}
