\name{print.kfhead}
\alias{print.kfhead}

\title{Print a kfhead object}
\description{
  Prints the header information from the \code{kftrack.dat} file.  
}
\usage{
print.kfhead(x, ...)
}
\arguments{
  \item{x}{is an \code{kfhead} object}
  \item{\dots}{additional arguments}
}
\author{John Sibert \email{jsibert@soest.hawaii.edu}, Anders Nielsen \email{anielsen@dina.kvl.dk}}
\seealso{\code{\link{print.kftrack}}}
\keyword{programming}
