\name{print.kftrack}
\alias{print.kftrack}
\title{Print kftrack object}
\description{
  Prints a pretty summary of an object of class \code{kftrack}
}
\usage{
print.kftrack(x, ...)
}
\arguments{
  \item{x}{an object of class \code{kftrack} typically generated
           with the \code{\link{kftrack}} function.}
  \item{\dots}{additional arguments}
}
\author{John Sibert \email{jsibert@soest.hawaii.edu}, Anders Nielsen \email{anielsen@dina.kvl.dk}}
\seealso{\code{\link{print.kftrack}}}
\keyword{programming}
